Added tests for encoder drive

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/main/utils/autonomous/location/pipeline/PositionSystem.java
@@ -78,7 +78,7 @@ public class PositionSystem {
         double y = northSouth.sensor1;
 
         // do some geometry-I honors level math
-        int angleDegrees = (int) coordinateSystem.angle.asDegree();
+        int angleDegrees = coordinateSystem.angle.asDegree();
         double angleRadians = Math.toRadians(angleDegrees);
 
         x = x * Math.sin(angleRadians);
